+++ src/domible/builders/__init__.py
@@ -3,6 +3,13 @@
 from domible.builders.tableBuilder import RowBuilder, TableBuilder, build_table_from_dicts
 from domible.builders.navBuilder import NavBuilder 
 from domible.builders.elementFromObject import element_from_object  
-from domible.builders.formBuilder import ToggleDetailsButton, default_toggle_details_button  
+from domible.builders.formBuilder import (
+    ToggleAllDetailsButton,
+    ExpandAllDetailsButton,
+    CollapseAllDetailsButton,
+    default_toggle_details_button,
+    default_expand_details_button,
+    default_collapse_details_button,
+)
 
 # end of file