+++ src/domible/builders/formBuilder.py
@@ -6,7 +6,7 @@ from typing import Any
 from domible.elements import Button, Div, Script 
 
 
-js_code = """
+js_toggle_all_code = """
 function toggleDetails(expand) {
     const detailsElements = document.querySelectorAll('details');
     detailsElements.forEach(details => {
@@ -25,25 +25,45 @@ function toggleAllDetails() {
 """
 
 
-class ToggleDetailsButton:
-    def __init__(self, contents: Any = None, **attributes):
-        """
-        generates a single purpose button to set the details 'opn' property to True/False 
-        attributes passed in are intended for the button element 
-        """
-        contents = "Toggle All Details" if not contents else contents 
-        self.toggle_button = Button(
-            contents, onclick="toggleAllDetails()", ** attributes
-        )
+class ToggleDetailsBase:
+    """
+    using this base, create buttons with different behavior by 
+    - passing in different onclick handlers
+    - setting contents to an appropriate label 
+    """
+
+    def __init__(self, contents: Any, **attributes):
+        self.toggle_button = Button(contents, ** attributes)
 
     def get_button(self):
-        return Div([Script(js_code), self.toggle_button])
+        return Div([
+            Script(js_toggle_all_code), 
+            self.toggle_button,
+        ])
 
     def __call__(self):
         return self.get_button()
 
+
+class ToggleAllDetailsButton(ToggleDetailsBase):
+    def __init__(self, contents: Any = "Toggle All Details", **kwArgs ):
+        super().__init__(contents=contents, onclick="toggleAllDetails()", **kwArgs)
+
+
+class ExpandAllDetailsButton(ToggleDetailsBase):
+    def __init__(self, contents: Any = "Expand All Details", **kwArgs):
+        super().__init__(contents=contents, onclick="toggleDetails(true)", **kwArgs)
+
+
+class CollapseAllDetailsButton(ToggleDetailsBase):
+    def __init__(self, contents: Any = "Collapse All Details", **kwArgs):
+        super().__init__(contents=contents, onclick="toggleDetails(false)", **kwArgs)
+
+
 ##
-# create this default to be easily used where the default label is acceptable 
-default_toggle_details_button = ToggleDetailsButton()
+# create this default to be easily used where the default label is acceptable
+default_toggle_details_button = ToggleAllDetailsButton()
+default_expand_details_button = ExpandAllDetailsButton()
+default_collapse_details_button = CollapseAllDetailsButton()
 
 ## end of file